+++ dlasset/export/functions/__init__.py
@@ -3,5 +3,5 @@ from .game_object import export_game_object
 from .image import export_image
 from .image_alpha import export_image_alpha
 from .image_story import export_image_story
-from .monobehaviour import export_mono_behaviour
+from .mono_behaviour import export_mono_behaviour
 from .sprite_ui import export_sprite_ui
